Refactoring code base so name is aligned with goals
