+++ llmcontextcompressor/config.py
@@ -1,0 +1,12 @@
+import os
+
+from dotenv import load_dotenv
+from pydantic import BaseSettings
+
+load_dotenv()
+
+
+class Settings(BaseSettings):
+    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY")
+
+settings = Settings()
