+++ llmcontextcompressor/constants.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class RankMethodType(Enum):
+    OPEN_AI = "OPEN_AI"
+    CLAUDE = "CLAUDE"
+    BM25 = "BM25"
