+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "llmpromptcompressor"
+name = "llmcontextcompressor"
 version = "0.1.0"
-description = "A Python Library used by LLM for Compressing Prompts"
+description = "A Python Library used by LLMs for Compressing Context"
 authors = ["Shannon Lal <shannonlal@hotmail.com>"]
 license = "Apache License"
 readme = "README.md"
@@ -12,13 +12,11 @@ python-dotenv = "^1.0.1"
 asyncio = "^3.4.3"
 aiohttp = "^3.9.4"
 tiktoken = "^0.6.0"
-sentence-transformers = "^2.6.1"
 numpy = "^1.26.4"
 openai = "^1.17.1"
 pydantic = "1.10.15"
 rank-bm25 = "^0.2.2"
 nltk = "^3.8.1"
-torch = "2.0.1"
 anthropic = "^0.25.6"
